requirements.txt
# Core ML packages (installed via conda)
# torch>=2.0.0
# torchvision>=0.15.0  
# torchaudio>=0.13.0

# Core packages
numpy>=1.24.0
scipy>=1.10.0
Pillow>=10.0.0
matplotlib>=3.7.0
packaging>=21.0

# Deep learning & diffusion models
diffusers>=0.21.0,<0.30.0
transformers>=4.30.0,<4.40.0
einops>=0.6.0

# Dataset loading (required for optim_utils.py)
datasets>=2.10.0
huggingface_hub>=0.15.0

# Image processing
torchvision>=0.15.0

# Cryptography
pycryptodome>=3.20.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
PyYAML>=6.0.0

# Optional packages for extended functionality
# datasets>=2.10.0        # For advanced dataset loading
# huggingface_hub>=0.15.0 # For model hub access
# timm>=0.9.0             # For additional vision models
# kornia>=0.6.0           # For geometric transformations  
# albumentations>=1.3.0   # For advanced augmentations
# scikit-image>=0.19.0    # For image processing
# opencv-python>=4.5.0    # For computer vision operations
